Move cube module from examples/linx_cpu_pyc/ to janus/pyc/janus/
- Relocated cube module to be alongside janus/bcc for better organization
- Updated all import paths from examples.linx_cpu_pyc.cube.* to janus.cube.*
- Updated README.md with new paths and references

The cube module now lives at janus/pyc/janus/cube/ alongside the
Janus BCC CPU implementation.
