--- examples/linx_cpu_pyc/cube/__init__.py
@@ -1,1 +1,0 @@
-# Cube matrix multiplication accelerator module
