+++ janus/pyc/janus/cube/__init__.py
@@ -1,1 +1,1 @@
-from __future__ import annotations
+# Cube matrix multiplication accelerator module
