+++ janus/pyc/janus/cube/cube_pe.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from pycircuit import Circuit, Wire, jit_inline
 
-from examples.linx_cpu_pyc.cube.cube_types import PERegs
+from janus.cube.cube_types import PERegs
 
 
 @jit_inline
